requirements.txt
beautifulsoup4>=4.12.0
fastapi>=0.109.0
google-auth>=2.27.0
mysql-connector-python>=8.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
selenium>=4.15.0
uvicorn[standard]>=0.27.0
lxml>=5.0.0
